fix rendering issue but it's currently in mapexpenses
