+++ src/Pages/MapExpenses.js
@@ -2,23 +2,24 @@ import "../App.css";
 import Map from "../Components/Map";
 import ListExpenses from "../Components/ListExpenses";
 import Welcome from "./Welcome";
-import { useState, useEffect } from "react";
+import { useState, useEffect, useMemo } from "react";
 import { realTimeDatabase } from "../firebase";
-import { ref, update, remove } from "firebase/database";
+import { ref, update, remove, off, onValue } from "firebase/database";
 import { useLoadScript } from "@react-google-maps/api";
 import { Toast } from "react-bootstrap";
 
-const DB_EXPENSES_FOLDER_NAME = "expenses";
 const DB_USER_FOLDER_NAME = "user";
+const DB_EXPENSES_FOLDER_NAME = "expenses";
+const DB_CATEGORY_FOLDER_NAME = "categories";
 
 export default function MapExpenses({
   isLoggedIn,
   uid,
   userData,
-  expensesCategory,
+  // expensesCategory,
   currenciesList,
-  categoriesData,
-  isLoading,
+  // categoriesData,
+  // isLoading,
 }) {
   const [userLocation, setUserLocation] = useState(null);
   const [isHighlighted, setIsHighlighted] = useState(null);
@@ -28,10 +29,112 @@ export default function MapExpenses({
   const [showToast, setShowToast] = useState(false);
   const [groupedExpenses, setGroupedExpenses] = useState([]);
   const [expenseCounter, setExpenseCounter] = useState(0);
+  const [expenses, setExpenses] = useState([]);
+  const [categoriesData, setCategoriesData] = useState([]);
+  // const [isLoading, setIsLoading] = useState(true);
+  const [isLoadingExpenses, setIsLoadingExpenses] = useState(false);
+  const [isLoadingCategories, setIsLoadingCategories] = useState(false);
+  const [expensesCategory, setExpensesCategory] = useState([]);
 
   // not used in current component
-  const [expRef, setExpRef] = useState();
   const [mapRef, setMapRef] = useState();
+
+  // Fetches latest category array, triggered with every change
+  useEffect(() => {
+    // setIsLoadingCategories(true);
+    const catRef = ref(realTimeDatabase, `${DB_CATEGORY_FOLDER_NAME}/${uid}`);
+    const unsubscribe = onValue(
+      catRef,
+      (snapshot) => {
+        const catData = snapshot.val();
+        // console.log(catData);
+        if (catData) {
+          const catArray = Object.entries(catData).map(([key, value]) => ({
+            id: key,
+            ...value,
+          }));
+          setCategoriesData((prevCategoriesData) =>
+            JSON.stringify(prevCategoriesData) !== JSON.stringify(catArray)
+              ? catArray
+              : prevCategoriesData
+          );
+        }
+        // setIsLoadingCategories(false); // <-- Set isLoadingCategories to false when fetch finishes
+      },
+      (errorObject) => {
+        console.log("The read failed: " + errorObject.name);
+        // setIsLoadingCategories(false); // <-- Also set isLoadingCategories to false in case of error
+      }
+    );
+
+    return () => {
+      // Remove the listener when the component unmounts
+      unsubscribe();
+    };
+  }, [uid]);
+  console.log("categoriesData:", categoriesData);
+
+  // Fetches latest expenses array, triggered with every change
+  useEffect(() => {
+    setIsLoadingExpenses(true); // <-- Set isLoadingExpenses to true when fetch starts
+
+    const expRef = ref(realTimeDatabase, `${DB_EXPENSES_FOLDER_NAME}/${uid}`);
+    const listener = onValue(
+      expRef,
+      (snapshot) => {
+        const expensesData = snapshot.val();
+        if (expensesData) {
+          const expensesArray = Object.entries(expensesData).map(
+            ([key, value]) => ({
+              id: key,
+              ...value,
+            })
+          );
+          // Sort expenses by date, with the latest at the top of the list
+          const sortedExpenses = expensesArray.sort(
+            (a, b) => new Date(b.date) - new Date(a.date)
+          );
+          setExpenses(sortedExpenses);
+          // Ensure that both expenses and categoriesData are loaded before attempting to join
+          if (!isLoadingCategories) {
+            const expensesCategory = sortedExpenses.map((expense, index) => {
+              const category = categoriesData.find(
+                (category) => category.category === expense.categoryName
+              );
+              // Ensure a category is found. If not, provide a fallback category
+              const fallbackCategory = category
+                ? category
+                : { category: "Unknown", color: "#000000", emoji: "❓" };
+              // Modify the spread sequence so the id from expense is not overwritten.
+              return { ...fallbackCategory, ...expense };
+            });
+            setExpensesCategory(expensesCategory);
+
+            const groupedExpenses = {};
+            expensesCategory.forEach((expense) => {
+              const date = expense.date;
+              if (!groupedExpenses[date]) {
+                groupedExpenses[date] = [];
+              }
+              groupedExpenses[date].push(expense);
+            });
+            setGroupedExpenses(groupedExpenses);
+          }
+
+          setIsLoadingExpenses(false);
+          console.log("expenses", expenses);
+        }
+      },
+      (error) => {
+        console.error(error);
+        setIsLoadingExpenses(false); // <-- Also set isLoadingExpenses to false in case of error
+      }
+    );
+    return () => {
+      off(expRef, listener);
+      setExpenses([]);
+    };
+  }, [uid, isLoadingCategories, categoriesData]);
 
   // Get user's location and assign coordinates to states
   useEffect(() => {
@@ -56,18 +159,20 @@ export default function MapExpenses({
   }, [expenseCounter]);
   // console.log("expenseCounter", expenseCounter);
 
-  // Group expenses with category by date
-  useEffect(() => {
-    const groupedExpenses = {};
-    expensesCategory.forEach((expense) => {
-      const date = expense.date;
-      if (!groupedExpenses[date]) {
-        groupedExpenses[date] = [];
-      }
-      groupedExpenses[date].push(expense);
-    });
-    setGroupedExpenses(groupedExpenses);
-  }, [expensesCategory]);
+  // // Group expenses with category by date
+  // useEffect(() => {
+  //   if (expensesCategory.length > 0) {
+  //     const groupedExpenses = {};
+  //     expensesCategory.forEach((expense) => {
+  //       const date = expense.date;
+  //       if (!groupedExpenses[date]) {
+  //         groupedExpenses[date] = [];
+  //       }
+  //       groupedExpenses[date].push(expense);
+  //     });
+  //     setGroupedExpenses(groupedExpenses);
+  //   }
+  // }, [expensesCategory]);
   // console.log("Grouped expenses:", groupedExpenses);
 
   // Fetches displayCurrency from the database and update the client-side state i.e. Database > Client
@@ -166,7 +271,8 @@ export default function MapExpenses({
             isHighlighted={isHighlighted}
             setIsHighlighted={setIsHighlighted}
             handleOnSelect={handleOnSelect}
-            isLoading={isLoading}
+            // isLoading={isLoading}
+            isLoadingExpenses={isLoadingExpenses}
             displayCurrency={displayCurrency}
             setDisplayCurrency={setDisplayCurrency}
             currenciesList={currenciesList}
